Filial Imprinting: hopefully fix static buff drops
Increasingly very concerned with Rivals RNG's level of bias and capacity to get seemingly stuck in periodic loops. Setting the seed variably based on gameplay time seems to help with breaking this looping issue.

Note also that setting random_func seeds based on the player is redundant, as the function's output is already player-dependent according to documentation.